Criando primeira ViewModel
